refactor: split large modules into slice directories
