+++ src/auth/mod.rs
@@ -1,0 +1,12 @@
+mod resolver;
+mod service_account;
+mod session;
+
+#[derive(Debug, Clone)]
+pub struct ResolvedAuth {
+    pub token: String,
+    pub source: &'static str,
+}
+
+pub use resolver::resolve_access_token;
+pub use session::validate_login_session_token;
